@@ -2,14 +2,16 @@ import React from "react";
 
 function Entry(props) {
   return (
-    <div className="term">
-      <dt>
-        <span className="emoji" role="img" aria-label="Tense Biceps">
+    <div className="term" data-atropos-offset="5" >
+      <div data-atropos-offset="-2">
+      <dt >
+        <span className="emoji" data-atropos-offset="3" role="img" aria-label="Tense Biceps">
           {props.emoji}
         </span>
         <span>{props.name}</span>
       </dt>
-      <dd>{props.description}</dd>
+      <dd >{props.description}</dd>
+      </div>
     </div>
   );
 }
