@@ -3,23 +3,476 @@ const emojipedia = [
     id: 1,
     emoji: "💪",
     name: "Tense Biceps",
-    meaning:
-      "“You can do that!” or “I feel strong!” Arm with tense biceps. Also used in connection with doing sports, e.g. at the gym."
-  },
-  {
+    meaning: "“You can do that!” or “I feel strong!” Arm with tense biceps. Also used in connection with doing sports, e.g. at the gym."
+    },
+    {
     id: 2,
     emoji: "🙏",
     name: "Person With Folded Hands",
-    meaning:
-      "Two hands pressed together. Is currently very introverted, saying a prayer, or hoping for enlightenment. Is also used as a “high five” or to say thank you."
-  },
-  {
+    meaning: "Two hands pressed together. Is currently very introverted, saying a prayer, or hoping for enlightenment. Is also used as a “high five” or to say thank you."
+    },
+    {
     id: 3,
     emoji: "🤣",
     name: "Rolling On The Floor, Laughing",
-    meaning:
-      "This is funny! A smiley face, rolling on the floor, laughing. The face is laughing boundlessly. The emoji version of “rofl“. Stands for „rolling on the floor, laughing“."
-  }
+    meaning: "This is funny! A smiley face, rolling on the floor, laughing. The face is laughing boundlessly. The emoji version of 'rofl'. Stands for 'rolling on the floor, laughing'."
+    },
+    {
+    id: 4,
+    emoji: "😊",
+    name: "Smiling Face with Smiling Eyes",
+    meaning: "A smiling face with smiling eyes and rosy cheeks. It shows happiness, contentment, or friendly intentions. Often used to express warm feelings, gratitude, or joy."
+    },
+    {
+    id: 5,
+    emoji: "🔥",
+    name: "Fire",
+    meaning: "A flame or fire emoji, used to represent a variety of literal and figurative fires. It can convey excitement, success, attractiveness, or intensity."
+    },
+    {
+    id: 6,
+    emoji: "❤️",
+    name: "Red Heart",
+    meaning: "A classic red heart emoji, used to express love, affection, or deep friendship. It can also convey gratitude, support, or be used to emphasize romantic feelings."
+    },
+    {
+    id: 7,
+    emoji: "🌟",
+    name: "Glowing Star",
+    meaning: "A yellow star with a radiant center. It represents various positive sentiments like excitement, accomplishment, recognition, or admiration."
+    },
+    {
+    id: 8,
+    emoji: "😂",
+    name: "Face with Tears of Joy",
+    meaning: "A laughing face with tears streaming down its cheeks. It indicates uncontrollable laughter or extreme amusement, often used to convey humor or joy."
+    },
+    {
+    id: 9,
+    emoji: "👍",
+    name: "Thumbs Up",
+    meaning: "A thumbs-up gesture indicating approval, agreement, or satisfaction. It can also be used to show support, encouragement, or a positive attitude."
+    },
+    {
+    id: 10,
+    emoji: "🤔",
+    name: "Thinking Face",
+    meaning: "A yellow face with a raised eyebrow, chin resting on its hand. It indicates thoughtfulness, contemplation, skepticism, or trying to come up with a solution."
+    },
+    {
+    id: 11,
+    emoji: "🎉",
+    name: "Party Popper",
+    meaning: "A party popper, an object that releases confetti and streamers to celebrate a special occasion. It is often used to convey congratulations or excitement."
+    },
+    {
+    id: 12,
+    emoji: "😍",
+    name: "Smiling Face with Heart-Eyes",
+    meaning: "A face with hearts instead of eyes, expressing deep affection, infatuation, or admiration. It is commonly used to convey love or adoration for someone or something."
+    },
+    {
+    id: 13,
+    emoji: "🎶",
+    name: "Musical Notes",
+    meaning: "Three eighth notes (quavers) connected by a beam, used to represent music or singing. It can indicate a love for music, a song or lyrics, or a joyful mood."
+    },
+    {
+    id: 14,
+    emoji: "🌸",
+    name: "Cherry Blossom",
+    meaning: "A pink flower, often depicted with a light-pink cherry blossom, representing the arrival of spring. It symbolizes beauty, nature, new beginnings, and feminine charm."
+    },
+    {
+    id: 15,
+    emoji: "👏",
+    name: "Clapping Hands",
+    meaning: "Two hands clapping together, used to applaud, praise, or show appreciation. It can also be used to encourage someone or as a gesture of congratulations."
+    },
+    {
+    id: 16,
+    emoji: "🙌",
+    name: "Raising Hands",
+    meaning: "Two hands raised in the air, celebrating or gesturing in a high-five. It can represent excitement, joy, praise, or a sense of accomplishment."
+    },
+    {
+    id: 17,
+    emoji: "🤷",
+    name: "Shrugging",
+    meaning: "A person shrugging their shoulders, indicating a lack of knowledge, confusion, indifference, or uncertainty. It can be used as a response to a question or situation."
+    },
+    {
+    id: 18,
+    emoji: "😎",
+    name: "Smiling Face with Sunglasses",
+    meaning: "A face with cool sunglasses, showing a confident, carefree, or detached attitude. It is often used to convey a sense of style or being cool."
+    },
+    {
+    id: 19,
+    emoji: "🌞",
+    name: "Sun with Face",
+    meaning: "A sun with a smiling face, often depicted with rays extending outward. It represents sunny weather, positivity, happiness, or a sense of warmth and comfort."
+    },
+    {
+    id: 20,
+    emoji: "🍕",
+    name: "Pizza",
+    meaning: "A slice of pizza, with cheese, tomato sauce, and toppings on a round, golden-brown crust. It symbolizes pizza, food cravings, or a casual meal."
+    },
+    {
+    id: 21,
+    emoji: "🥰",
+    name: "Smiling Face with Hearts",
+    meaning: "A face with smiling eyes and three hearts floating in the air, expressing extreme love, adoration, or feeling touched by a heartfelt gesture."
+    },
+    {
+    id: 22,
+    emoji: "😊",
+    name: "Smiling Face with Smiling Eyes",
+    meaning: "A smiling face with smiling eyes and rosy cheeks. It shows happiness, contentment, or friendly intentions. Often used to express warm feelings, gratitude, or joy."
+    },
+    {
+    id: 23,
+    emoji: "🐶",
+    name: "Dog Face",
+    meaning: "The face of a dog, which may be depicted as a variety of breeds. It represents dogs, pets, loyalty, companionship, or a love for animals."
+    },
+    {
+    id: 24,
+    emoji: "😘",
+    name: "Face Blowing a Kiss",
+    meaning: "A face with pursed lips sending a kiss to someone or something. It conveys affection, love, or gratitude and can be used romantically or in a friendly manner."
+    },
+    {
+    id: 25,
+    emoji: "🌈",
+    name: "Rainbow",
+    meaning: "A colorful arc-shaped spectrum, showing the seven colors of the rainbow. It symbolizes hope, diversity, unity, happiness, or the LGBTQ+ pride movement."
+    },
+    {
+    id: 26,
+    emoji: "🎈",
+    name: "Balloon",
+    meaning: "A colorful balloon on a string, used for decoration and celebration. It represents joy, fun, parties, birthdays, or a sense of childlike wonder."
+    },
+    {
+    id: 27,
+    emoji: "🍀",
+    name: "Four Leaf Clover",
+    meaning: "A four-leaf clover, a universally recognized symbol of good luck. It represents fortune, luck, hope, and the pursuit of serendipity or unexpected surprises."
+    },
+    {
+    id: 28,
+    emoji: "😉",
+    name: "Winking Face",
+    meaning: "A face with one eye closed and a slight smile. It indicates humor, mischief, playfulness, or a shared secret between the sender and recipient."
+    },
+    {
+    id: 29,
+    emoji: "👌",
+    name: "OK Hand",
+    meaning: "A hand gesture forming a circle by joining the thumb and index finger, with the remaining fingers raised. It signifies approval, agreement, or perfection."
+    },
+    {
+    id: 30,
+    emoji: "🌹",
+    name: "Rose",
+    meaning: "A red rose, a classic symbol of love and romance. It represents beauty, passion, affection, or can be used to convey admiration or a heartfelt gesture."
+    },
+    {
+    id: 31,
+    emoji: "😁",
+    name: "Grinning Face with Smiling Eyes",
+    meaning: "A face with a big, open smile and smiling eyes. It indicates happiness, joy, excitement, or a positive outlook on a situation."
+    },
+    {
+    id: 32,
+    emoji: "🎁",
+    name: "Wrapped Gift",
+    meaning: "A wrapped gift box with a bow on top, given for a special occasion like a birthday, Christmas, or celebration. It represents surprise, joy, or generosity."
+    },
+    {
+    id: 33,
+    emoji: "😢",
+    name: "Crying Face",
+    meaning: "A face with a single tear streaming down its cheek. It signifies sadness, disappointment, grief, or a deep sense of loss."
+    },
+    {
+    id: 34,
+    emoji: "🤗",
+    name: "Hugging Face",
+    meaning: "A face with open hands placed on the cheeks, expressing a warm, affectionate embrace. It can convey a hug, comfort, support, or excitement."
+    },
+    {
+    id: 35,
+    emoji: "🌻",
+    name: "Sunflower",
+    meaning: "A yellow sunflower, a tall plant with a large, round flower head. It symbolizes happiness, warmth, positivity, adoration, or the summer season."
+    },
+    {
+    id: 36,
+    emoji: "🐱",
+    name: "Cat Face",
+    meaning: "The face of a cat, which may be depicted as a variety of breeds. It represents cats, pets, independence, playfulness, or a love for feline creatures."
+    },
+    {
+    id: 37,
+    emoji: "😭",
+    name: "Loudly Crying Face",
+    meaning: "A face with streaming tears and a wide-open mouth, crying loudly. It expresses intense sadness, grief, pain, or a strong emotional reaction."
+    },
+    {
+    id: 38,
+    emoji: "🥳",
+    name: "Partying Face",
+    meaning: "A face wearing a party hat and blowing a party horn, often depicted with confetti and streamers. It represents celebration, excitement, or joyous occasions."
+    },
+    {
+    id: 39,
+    emoji: "✨",
+    name: "Sparkles",
+    meaning: "A cluster of glittering sparkles, often used to represent various magic, festive, or joyful themes. It can also convey cleanliness or a touch of glamour."
+    },
+    {
+    id: 40,
+    emoji: "🌺",
+    name: "Hibiscus",
+    meaning: "A red hibiscus flower, often depicted as a single blossom with prominent petals. It symbolizes beauty, femininity, tropical settings, or a love for nature."
+    },
+    {
+      id: 41,
+      emoji: "🙃",
+      name: "Upside-Down Face",
+      meaning: "A face rotated 180 degrees, with the eyes and mouth positioned upside down. It can indicate sarcasm, irony, or a sense of joking or playfulness."
+      },
+      {
+      id: 42,
+      emoji: "🍎",
+      name: "Red Apple",
+      meaning: "A red apple, representing the fruit itself, health, nutrition, or teacher appreciation. It can also symbolize knowledge, learning, or a healthy lifestyle."
+      },
+      {
+      id: 43,
+      emoji: "🤩",
+      name: "Star-Struck",
+      meaning: "A face with starry eyes and a wide grin, expressing extreme excitement, admiration, or amazement. It is often used to show overwhelming approval or love."
+      },
+      {
+      id: 44,
+      emoji: "🐾",
+      name: "Paw Prints",
+      meaning: "The marks left by the paws of an animal, often used to represent dogs, cats, or other pets. It can symbolize animal tracks, loyalty, or a love for animals."
+      },
+      {
+      id: 45,
+      emoji: "😴",
+      name: "Sleeping Face",
+      meaning: "A face with eyes closed, indicating sleep, rest, or tiredness. It can also convey boredom, disinterest, or a lack of enthusiasm."
+      },
+      {
+      id: 46,
+      emoji: "🌊",
+      name: "Water Wave",
+      meaning: "A stylized representation of a wave, often depicted in blue. It symbolizes water, the ocean, beach, relaxation, or a sense of calmness and tranquility."
+      },
+      {
+      id: 47,
+      emoji: "🎵",
+      name: "Musical Note",
+      meaning: "A single eighth note (quaver), used to represent music or singing. It can indicate a love for music, a song or melody, or an interest in musical activities."
+      },
+      {
+      id: 48,
+      emoji: "😃",
+      name: "Grinning Face with Big Eyes",
+      meaning: "A face with a big, open smile and wide, gleaming eyes. It conveys happiness, joy, excitement, or a positive outlook on a situation."
+      },
+      {
+      id: 49,
+      emoji: "🎉",
+      name: "Party Popper",
+      meaning: "A cone-shaped party popper, bursting with confetti and streamers. It represents celebration, excitement, joy, or a festive atmosphere."
+      },
+      {
+      id: 50,
+      emoji: "😄",
+      name: "Grinning Face with Smiling Eyes",
+      meaning: "A face with a big, open smile and smiling eyes. It indicates happiness, joy, excitement, or a positive outlook on a situation."
+      },
+      {
+      id: 51,
+      emoji: "🌟",
+      name: "Glowing Star",
+      meaning: "A stylized, radiant star, often depicted in yellow or gold. It symbolizes achievement, success, brilliance, good luck, or a special occasion."
+      },
+      {
+      id: 52,
+      emoji: "🔥",
+      name: "Fire",
+      meaning: "A flame, representing fire itself. It can symbolize heat, passion, intensity, energy, transformation, or a sense of excitement and enthusiasm."
+      },
+      {
+      id: 53,
+      emoji: "🥺",
+      name: "Pleading Face",
+      meaning: "A face with wide, shimmering eyes and a slight frown, expressing a pleading or begging emotion. It can convey a request, apology, or puppy dog eyes."
+      },
+      {
+      id: 54,
+      emoji: "🌙",
+      name: "Crescent Moon",
+      meaning: "A crescent moon, representing the moon during its waxing or waning phase. It symbolizes nighttime, dreams, magic, mystery, or a sense of calmness."
+      },
+      {
+      id: 55,
+      emoji: "😍",
+      name: "Smiling Face with Heart-Eyes",
+      meaning: "A face with smiling eyes and hearts instead of eyes, expressing deep affection, infatuation, or admiration. It conveys love, adoration, or being smitten."
+      },
+      {
+      id: 56,
+      emoji: "🌍",
+      name: "Globe Showing Europe-Africa",
+      meaning: "A globe showing a portion of Europe and Africa. It represents the Earth, global awareness, travel, international connections, or a concern for the world."
+      },
+      {
+      id: 57,
+      emoji: "😊",
+      name: "Smiling Face with Smiling Eyes",
+      meaning: "A face with a big, open smile, smiling eyes, and rosy cheeks. It shows happiness, contentment, friendliness, or a warm and welcoming demeanor."
+      },
+      {
+      id: 58,
+      emoji: "🌸",
+      name: "Cherry Blossom",
+      meaning: "A pink cherry blossom flower, typically associated with Japanese culture and the arrival of spring. It symbolizes beauty, renewal, or fleeting nature."
+      },
+      {
+      id: 59,
+      emoji: "🤔",
+      name: "Thinking Face",
+      meaning: "A face with a raised eyebrow and a finger resting on the chin, deep in thought. It conveys contemplation, reflection, or considering various possibilities."
+      },
+      {
+      id: 60,
+      emoji: "🥰",
+      name: "Smiling Face with Hearts",
+      meaning: "A face with smiling eyes and three hearts floating above it, expressing extreme affection, love, or adoration. It conveys overwhelming happiness and love."
+      },
+      {
+      id: 61,
+      emoji: "🎊",
+      name: "Confetti Ball",
+      meaning: "A round confetti ball, filled with colorful confetti paper. It represents celebration, parties, joy, festivity, or a festive atmosphere."
+      },
+      {
+      id: 62,
+      emoji: "🌞",
+      name: "Sun with Face",
+      meaning: "A sun with a smiling face, often depicted with rays of sunshine. It symbolizes warmth, happiness, positivity, or a bright and sunny day."
+      },
+      {
+      id: 63,
+      emoji: "😑",
+      name: "Expressionless Face",
+      meaning: "A face with a flat, neutral expression and closed mouth. It conveys a lack of emotion, indifference, boredom, or an uninterested attitude."
+      },
+      {
+      id: 64,
+      emoji: "😋",
+      name: "Face Savoring Food",
+      meaning: "A face with a savoring or licking expression, often associated with delicious food or a pleasurable eating experience. It conveys enjoyment and satisfaction."
+      },
+      {
+      id: 65,
+      emoji: "🌴",
+      name: "Palm Tree",
+      meaning: "A tall palm tree with a single trunk and a cluster of leaves at the top. It symbolizes tropical settings, vacations, relaxation, or a love for nature."
+      },
+      {
+      id: 66,
+      emoji: "🤣",
+      name: "Rolling On The Floor, Laughing",
+      meaning: "This is funny! A smiley face, rolling on the floor, laughing. The face is laughing boundlessly. The emoji version of 'rofl.' Stands for 'rolling on the floor, laughing.'"
+      },
+      {
+      id: 67,
+      emoji: "🌟",
+      name: "Glowing Star",
+      meaning: "A stylized, radiant star, often depicted in yellow or gold. It symbolizes achievement, success, brilliance, good luck, or a special occasion."
+      },
+      {
+      id: 68,
+      emoji: "🍕",
+      name: "Slice of Pizza",
+      meaning: "A slice of pizza, topped with cheese, tomato sauce, and various toppings. It represents pizza, fast food, indulgence, or a craving for something tasty."
+      },
+      {
+        id: 69,
+        emoji: "💔",
+        name: "Broken Heart",
+        meaning: "A red heart split in two, representing a broken or shattered heart. It symbolizes heartbreak, sadness, disappointment, or a lost love."
+        },
+        {
+        id: 70,
+        emoji: "🌈",
+        name: "Rainbow",
+        meaning: "A colorful arc-shaped spectrum, representing the optical phenomenon of a rainbow. It symbolizes diversity, inclusivity, hope, or a positive outlook."
+        },
+        {
+        id: 71,
+        emoji: "🍺",
+        name: "Beer Mug",
+        meaning: "A frothy beer mug, often depicted with golden beer overflowing from the top. It represents beer, alcohol, celebrations, or a social gathering."
+        },
+        {
+        id: 72,
+        emoji: "🙌",
+        name: "Raising Hands",
+        meaning: "Two hands raised in celebration or excitement, often depicted with palms facing upward. It symbolizes praise, applause, joy, or a victorious moment."
+        },
+        {
+        id: 73,
+        emoji: "😡",
+        name: "Pouting Face",
+        meaning: "A face with a frowning mouth, furrowed eyebrows, and a pouting expression. It conveys anger, frustration, annoyance, or a sense of displeasure."
+        },
+        {
+        id: 74,
+        emoji: "🎁",
+        name: "Wrapped Gift",
+        meaning: "A wrapped gift box, often with a bow on top. It represents a present, gift-giving, surprises, celebrations, or a special occasion."
+        },
+        {
+        id: 75,
+        emoji: "🤷",
+        name: "Person Shrugging",
+        meaning: "A person with raised shoulders and upturned hands, representing a gesture of uncertainty, indifference, or a lack of knowledge or understanding."
+        },
+        {
+        id: 76,
+        emoji: "🌹",
+        name: "Rose",
+        meaning: "A red rose, typically depicted as a single long-stemmed flower with thorns. It symbolizes love, romance, beauty, passion, or a special someone."
+        },
+        {
+        id: 77,
+        emoji: "😅",
+        name: "Grinning Face with Sweat",
+        meaning: "A face with a big, open smile and bead of sweat on the forehead, representing nervousness, embarrassment, or a sense of relief after hard work."
+        },
+        {
+        id: 78,
+        emoji: "🎶",
+        name: "Musical Notes",
+        meaning: "A set of musical notes, typically depicted as two or more eighth notes (quavers). It represents music, singing, melody, or a love for musical activities."
+        },
+        {
+        id: 79,
+        emoji: "🥳",
+        name: "Partying Face",
+        meaning: "A face wearing a party hat and blowing a party horn, often depicted with confetti and streamers. It represents celebration, excitement, or joyous occasions."
+        },
 ];
 
 export default emojipedia;
