@@ -1,4 +1,4 @@
-export const data = {
+export const data = [{
   id: "10000",
   slug: "test-product",
   product_name: "Test product",
@@ -9,4 +9,53 @@ export const data = {
     color: ["red", "black", "blue", "white"],
     size: ["us-5.5", "us-6", "us-6.5", "eu-36", "eu-36.5", "eu-37"],
   },
-};
+},
+{
+  id: "10001",
+  slug: "test-product",
+  product_name: "Test product",
+  product_image:
+    "https://image.freepik.com/free-psd/product-sale-social-media-post-banner-template_135461-151.jpg",
+  product_price: 10000,
+  variations: {
+    color: ["red", "black", "blue", "white"],
+    size: ["us-5.5", "us-6", "us-6.5", "eu-36", "eu-36.5", "eu-37"],
+  },
+},
+{
+  id: "10002",
+  slug: "test-product",
+  product_name: "Test product",
+  product_image:
+    "https://media.istockphoto.com/id/1148822167/photo/close-up-shot-of-runners-shoes.jpg?s=612x612&w=0&k=20&c=jBfVGW1Ot-FXdA2AQ2eXVrKYcf3BsLGbkYYnRi2K_TE=",
+  product_price: 10000,
+  variations: {
+    color: ["red", "black", "blue", "white"],
+    size: ["us-5.5", "us-6", "us-6.5", "eu-36", "eu-36.5", "eu-37"],
+  },
+},
+{
+  id: "10003",
+  slug: "test-product",
+  product_name: "Test product",
+  product_image:
+    "https://hips.hearstapps.com/hmg-prod/images/run-affordable-running-shoes-1676481206.jpg?crop=0.502xw:1.00xh;0.250xw,0&resize=1200:*",
+  product_price: 10000,
+  variations: {
+    color: ["red", "black", "blue", "white"],
+    size: ["us-5.5", "us-6", "us-6.5", "eu-36", "eu-36.5", "eu-37"],
+  },
+},
+{
+  id: "10004",
+  slug: "test-product",
+  product_name: "Test product",
+  product_image:
+    "https://images.bewakoof.com/t640/men-s-white-casual-shoes-592552-1684502213-1.JPG",
+  product_price: 10000,
+  variations: {
+    color: ["red", "black", "blue", "white"],
+    size: ["us-5.5", "us-6", "us-6.5", "eu-36", "eu-36.5", "eu-37"],
+  },
+},
+]
